post list merged with post component and user can click post-creator name to navigate to their profile
