+++ web/src/graphql/createPost.graphql
@@ -7,6 +7,7 @@ mutation CreatePost($creatorId: Float!, $content: String!) {
       id
       firstName
       lastName
+      email
     }
     likes {
       user {
