+++ web/src/graphql/getUserPosts.graphql
@@ -17,6 +17,7 @@ query GetUserPosts(
       id
       firstName
       lastName
+      email
     }
     likes {
       user {
