+++ web/src/graphql/likePost.graphql
@@ -7,6 +7,7 @@ mutation LikePost($userId: Float!, $postId: Float!, $isAlreadyLiked: Boolean!) {
       id
       firstName
       lastName
+      email
     }
     likes {
       user {
